+++ backend-lib/common-securities/src/main/java/com/optimize/common/securities/dto/RequirePasswordChangeDto.java
@@ -1,0 +1,12 @@
+package com.optimize.common.securities.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class RequirePasswordChangeDto {
+    @NotEmpty(message = "La liste des utilisateurs est obligatoire.")
+    private List<Long> userIds;
+}